v1.5.30 -- After many tests, I ended up using a Makefile so the bumphunter example output will be included in the pkg
git-svn-id: file:///home/git/hedgehog.fhcrc.org/bioconductor/trunk/madman/Rpacks/regionReport@115595 bc3139a8-67e5-0310-9ffc-ced21a209358

+++ R/DESeq2Report.R
@@ -135,10 +135,13 @@ DESeq2Report <- function(dds, project = "", intgroup, colors = NULL, res = NULL,
     
     ## Locate Rmd if one is not provided
     if (is.null(template)) {
+        templateNull <- TRUE
         template <- system.file(
             file.path('DESeq2Exploration', 'DESeq2Exploration.Rmd'),
             package = 'regionReport', mustWork = TRUE
         )
+    } else {
+        templateNull <- FALSE
     }
     
     ## Load knitcitations with a clean bibliography
@@ -199,7 +202,7 @@ DESeq2Report <- function(dds, project = "", intgroup, colors = NULL, res = NULL,
             res <- render(paste0(output, '.Rmd'), output_format,
                 clean = .advanced_argument('clean', TRUE, ...))
         }
-        file.remove(paste0(output, '.Rmd'))
+        if(templateNull) file.remove(paste0(output, '.Rmd'))
     
         ## Open
         if (browse) browseURL(res)
